# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HabitLab API",
    "version": "1.0.0",
    "description": "API REST de gestion d'habitudes. Toutes les routes métier exigent `Authorization: Bearer <JWT>` ; chaque compte ne voit que ses propres habitudes."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "INVALID_INPUT",
                  "UNAUTHORIZED",
                  "NOT_FOUND",
                  "EMAIL_ALREADY_USED"
                ]
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "Habit": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "507f191e810c19729de860ea"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120,
            "example": "Marcher"
          },
          "frequency": {
            "type": "string",
            "enum": [
              "daily",
              "weekly"
            ]
          },
          "active": {
            "type": "boolean"
          },
          "color": {
            "type": "string",
            "example": "#7cb73b",
            "description": "Extra : couleur d'affichage"
          },
          "completions": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "2026-10-07"
            },
            "description": "Extra : jours validés (YYYY-MM-DD)"
          },
          "ownerId": {
            "type": "string"
          }
        }
      },
      "HabitCreate": {
        "type": "object",
        "required": [
          "title",
          "frequency",
          "active"
        ],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "frequency": {
            "type": "string",
            "enum": [
              "daily",
              "weekly"
            ]
          },
          "active": {
            "type": "boolean"
          },
          "color": {
            "type": "string",
            "example": "#7cb73b"
          }
        },
        "description": "id et ownerId sont interdits (400)."
      },
...
    }
  },
  "paths": {
...
    "/api/habits": {
      "get": {
        "tags": [
          "Habits"
        ],
        "summary": "Lister mes habitudes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Liste (vide : items = [])",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Habit"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Jeton absent, invalide ou expiré (UNAUTHORIZED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Habits"
        ],
        "summary": "Créer une habitude",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HabitCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Habitude créée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            }
          },
          "400": {
            "description": "Corps invalide ou identifiant malformé (INVALID_INPUT)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Jeton absent, invalide ou expiré (UNAUTHORIZED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}